backend/requirements.txt
fastapi
uvicorn[standard]
paho-mqtt
pydantic
python-multipart
influxdb-client
httpx
